RE-57: Refactor
