+++ internal/auth-service/controller/http/auth/registerConfirmHandler.go
@@ -3,8 +3,8 @@ package auth
 import (
 	"encoding/json"
 	"net/http"
-	entity "pkg/entity"
-	"pkg/utils/validator"
+	entity "retarget/pkg/entity"
+	"retarget/pkg/utils/validator"
 )
 
 type RegisterConfirmRequest struct {
