--- internal/pkg/go.mod
@@ -1,22 +1,0 @@
-module pkg
-
-go 1.24.0
-
-require (
-	github.com/go-playground/validator/v10 v10.26.0
-	github.com/redis/go-redis/v9 v9.7.3
-	github.com/sirupsen/logrus v1.9.3
-)
-
-require (
-	github.com/cespare/xxhash/v2 v2.2.0 // indirect
-	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
-	github.com/gabriel-vasile/mimetype v1.4.8 // indirect
-	github.com/go-playground/locales v0.14.1 // indirect
-	github.com/go-playground/universal-translator v0.18.1 // indirect
-	github.com/leodido/go-urn v1.4.0 // indirect
-	golang.org/x/crypto v0.33.0 // indirect
-	golang.org/x/net v0.34.0 // indirect
-	golang.org/x/sys v0.30.0 // indirect
-	golang.org/x/text v0.22.0 // indirect
-)
