+++ pkg/middleware/auth/checkAuth.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"pkg/entity"
+	"retarget/pkg/entity"
 )
 
 func AuthMiddleware(authenticator AuthenticatorInterface) func(http.Handler) http.Handler {
